+++ src/main/java/monday/exception/UnknownCommandException.java
@@ -1,0 +1,11 @@
+package monday.exception;
+
+/**
+ * Thrown when the user enters a command that is not recognized by the system.
+ * Helps guide users toward valid commands.
+ */
+public class UnknownCommandException extends Exception {
+    public UnknownCommandException() {
+        super("I'm sorry, but I don't recognize that command. Please try again.");
+    }
+}
