+++ src/main/java/monday/parser/Parser.java
@@ -1,3 +1,17 @@
+package monday.parser;
+
+import monday.exception.EmptyDescriptionException;
+import monday.exception.InvalidCommandFormatException;
+import monday.exception.InvalidDateTimeException;
+import monday.exception.InvalidTaskNumberException;
+import monday.exception.UnknownCommandException;
+import monday.task.Task;
+import monday.task.TaskList;
+import monday.task.Todo;
+import monday.task.Deadline;
+import monday.task.Event;
+import monday.ui.Ui;
+
 /**
  * Handles parsing and interpretation of user commands.
  * Follows Single Responsibility Principle - only handles command parsing operations.
