+++ src/main/java/monday/storage/Storage.java
@@ -1,3 +1,5 @@
+package monday.storage;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.FileNotFoundException;
@@ -6,6 +8,12 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Scanner;
+
+import monday.exception.TaskLoadingException;
+import monday.task.Task;
+import monday.task.Todo;
+import monday.task.Deadline;
+import monday.task.Event;
 
 /**
  * Handles file storage operations for task persistence.
