+++ src/main/java/monday/task/Todo.java
@@ -1,4 +1,6 @@
-class Todo extends Task {
+package monday.task;
+
+public class Todo extends Task {
     public Todo(String description) {
         super(description);
     }
